+++ App.tsx
@@ -775,7 +775,7 @@ export const App: React.FC<IProps> = (): React.ReactElement => {
                   items={[
                     {
                       label: 'Doggo 1',
-                      jsx: (
+                      jsxLeftElement: (
                         <View
                           // eslint-disable-next-line react-native/no-inline-styles
                           style={{flexDirection: 'row', alignItems: 'center'}}>
@@ -796,7 +796,7 @@ export const App: React.FC<IProps> = (): React.ReactElement => {
                     },
                     {
                       label: 'Doggo 2',
-                      jsx: (
+                      jsxLeftElement: (
                         <View
                           // eslint-disable-next-line react-native/no-inline-styles
                           style={{flexDirection: 'row', alignItems: 'center'}}>
@@ -814,6 +814,99 @@ export const App: React.FC<IProps> = (): React.ReactElement => {
                     {
                       label: 'Simple Value',
                       value: 'Lookie Here Again!',
+                    },
+                  ]}
+                />
+              </Row>
+              <Row>
+                <DataBlock
+                  items={[
+                    {
+                      jsxLeftElement: (
+                        <View
+                          // eslint-disable-next-line react-native/no-inline-styles
+                          style={{flexDirection: 'row', alignItems: 'center'}}>
+                          <View>
+                            <Avatar uri={avatarUri} radius="full" />
+                          </View>
+                          <View
+                            // eslint-disable-next-line react-native/no-inline-styles
+                            style={{marginLeft: 10, flex: 1}}>
+                            <Heading variant="h4">Unclaimed</Heading>
+                            <View
+                              // eslint-disable-next-line react-native/no-inline-styles
+                              style={{
+                                flexDirection: 'row',
+                                alignItems: 'center',
+                                paddingVertical: minorScale(0.5),
+                              }}>
+                              <Text
+                                numberOfLines={1}
+                                variant={'subtitle2'}
+                                color={'graySix'}>
+                                5 Active
+                                <Text
+                                  numberOfLines={1}
+                                  weight={'semiBold'}
+                                  variant={'subtitle2'}
+                                  color={'warningLight'}>
+                                  (2 Overdue)
+                                </Text>
+                              </Text>
+                            </View>
+                          </View>
+                        </View>
+                      ),
+                    },
+                  ]}
+                />
+                <DataBlock
+                  items={[
+                    {
+                      jsxLeftElement: (
+                        <View
+                          // eslint-disable-next-line react-native/no-inline-styles
+                          style={{flexDirection: 'row', alignItems: 'center'}}>
+                          <View>
+                            <Avatar uri={avatarUri} radius="full" />
+                          </View>
+                          <View
+                            // eslint-disable-next-line react-native/no-inline-styles
+                            style={{marginLeft: 10, flex: 1}}>
+                            <Heading variant="h4">Team Member 1</Heading>
+                            <View
+                              // eslint-disable-next-line react-native/no-inline-styles
+                              style={{
+                                flexDirection: 'row',
+                                alignItems: 'center',
+                                paddingVertical: minorScale(0.5),
+                              }}>
+                              <Text
+                                numberOfLines={1}
+                                variant={'subtitle2'}
+                                color={'graySix'}>
+                                5/10 Complated
+                                <Text
+                                  numberOfLines={1}
+                                  weight={'semiBold'}
+                                  variant={'subtitle2'}
+                                  color={'warningLight'}>
+                                  (2 Overdue)
+                                </Text>
+                              </Text>
+                            </View>
+                          </View>
+                        </View>
+                      ),
+                      jsxRightElement: (
+                        <View
+                          // eslint-disable-next-line react-native/no-inline-styles
+                          style={{flexDirection: 'row', alignItems: 'center'}}>
+                          <View>
+                            <Avatar uri={avatarUri} radius="full" />
+                          </View>
+                        </View>
+                      ),
                     },
                   ]}
                 />