+++ src/modules/data-block/data-block.tsx
@@ -7,7 +7,7 @@ import {IStyledRow, IDataBlockItem, IProps} from './data-block.types';
 
 // Shared
 import Text from '../text/text';
-import {majorScale} from '../scales';
+import {minorScale} from '../scales';
 
 // Styles
 export const StyledWrapper = Styled.View`
@@ -18,23 +18,26 @@ export const StyledRow = Styled.View<IStyledRow>`
   flex-direction: row;
   justify-content: space-between;
   align-items: center;
-
-  padding-vertical: ${majorScale(1)}px;
+  padding-vertical: ${minorScale(3)}px;
   border-bottom-width: ${({$hideBorder}) => ($hideBorder ? 0 : 1)}px;
   border-bottom-color: ${({theme}) => theme.colors.grayThree};
 `;
 export const StyledLeftWrapper = Styled.View`
-  flex: 0.6;
+  flex: 1;
   align-items: flex-start;
 `;
 export const StyledRightWrapper = Styled.View`
-  flex: 0.4;
   align-items: flex-end;
+  flex: 0.35;
+`;
+export const StyledRightTextWrapper = Styled(Text)`
+  font-size: 12px;
 `;
 
 // Component
 export const DataBlock: React.FC<IProps> = ({
   items,
+  hideBottomLine,
 }): React.ReactElement<any, any> | null => {
   if (!items) {
     return null;
@@ -48,19 +51,23 @@ export const DataBlock: React.FC<IProps> = ({
         const isLastRow = index === items.length - 1;
 
         return (
-          <StyledRow key={key} $hideBorder={isLastRow}>
+          <StyledRow key={key} $hideBorder={hideBottomLine ? true : isLastRow}>
             <StyledLeftWrapper>
-              {item.jsx ? (
-                item.jsx
+              {item.jsxLeftElement ? (
+                item.jsxLeftElement
               ) : (
                 <Text variant="caption">{item.value}</Text>
               )}
             </StyledLeftWrapper>
 
             <StyledRightWrapper>
-              <Text variant="caption" muted>
-                {item.label}
-              </Text>
+              {item.jsxRightElement ? (
+                item.jsxRightElement
+              ) : (
+                <StyledRightTextWrapper muted>
+                  {item.label}
+                </StyledRightTextWrapper>
+              )}
             </StyledRightWrapper>
           </StyledRow>
         );